Add files via upload
This is the solution of the titanic problem statement on kaggle. This was my first contest submission on kaggle. I used logistic regression to predict the passengers who survived on the basis of age, fare of ticket, sex and the place of embarkment.  The kaggle score for this project was 0.76315